Created the first commit
RoulettePie comes up with thousands of random strategies to try to beat Roulette. All strategies are iterated around doubling your even bet (on red) each time you lose. It takes a strategy, tests it over 10 iterations, if it passes (ie. it doesn't break the bank) it tests to 100, again if it passes, it then tests to 1000, and finally 5000 iterations. After finding 10 strategies which pass 5000 iterations, it tests each one indefinitely until the bank runs out.  It has yet to find a winning strategy.  Remember - this is looking for a staking strategy rather than 'a way to predict what will happen'.... I think there are improvements which can be made.